requirements.txt
Flask==3.0.3
firebase-admin==6.4.0
flask-jwt-extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Login>=0.6.0
fpdf2==2.7.9
Werkzeug==3.0.3
openpyxl==3.1.2
playwright>=1.40
gunicorn>=21.0
boto3>=1.26
psycopg2-binary>=2.9
ezdxf>=1.1
PyMuPDF>=1.23
opencv-python-headless>=4.8
numpy>=1.24
